Restructure; Clean up; Add tests.
