+++ lib/CommandPath.ts
@@ -1,0 +1,3 @@
+// A command path is a list of command names and represents a path through a
+// cli's command hierarchy.
+export type CommandPath = string[];
