--- lib/ShowUsageFn.ts
@@ -1,3 +1,0 @@
-export type ShowUsageFn = (params: {
-  commandPath: string[];
-}) => void;
