+++ lib/commands/HelpOptions.ts
@@ -1,0 +1,5 @@
+import { CommandPath } from '../CommandPath';
+
+export interface HelpOptions {
+  command: undefined | CommandPath;
+}
