+++ lib/index.ts
@@ -1,9 +1,13 @@
-import { Command } from './Command';
 import { OptionDefinition } from './OptionDefinition';
 import { runCli } from './runCli';
+import { ShowUsageFn } from './usage/ShowUsageFn';
+import { Command, CommandHandle, CommandHandleOptions } from './Command';
 
 export {
   Command,
+  CommandHandle,
+  CommandHandleOptions,
   OptionDefinition,
-  runCli
+  runCli,
+  ShowUsageFn
 };
