+++ lib/recommend/RecommendCommandFn.ts
@@ -1,0 +1,5 @@
+import { CommandPath } from '../CommandPath';
+
+export type RecommendCommandFn = (params: {
+  commandPath: CommandPath;
+}) => CommandPath[];
