+++ lib/usage/ShowUsageFn.ts
@@ -1,0 +1,5 @@
+import { CommandPath } from '../CommandPath';
+
+export type ShowUsageFn = (params: {
+  commandPath: CommandPath;
+}) => string;
