+++ package.json
@@ -23,6 +23,8 @@
     "@types/command-line-args": "5.0.0",
     "@types/command-line-commands": "2.0.1",
     "@types/command-line-usage": "5.0.1",
+    "assertthat": "5.1.0",
+    "record-stdstreams": "3.0.0",
     "roboter": "10.0.5",
     "semantic-release-configuration": "1.0.6"
   },
